+++ payments-core/src/test/java/com/stripe/android/PaymentMethodEndToEndTest.kt
@@ -484,6 +484,15 @@ internal class PaymentMethodEndToEndTest {
     }
 
     @Test
+    fun createPaymentMethod_withSunbit_shouldCreateObject() {
+        val params = PaymentMethodCreateParamsFixtures.SUNBIT
+        val stripe = Stripe(context, ApiKeyFixtures.SUNBIT_PUBLISHABLE_KEY)
+
+        val paymentMethod = stripe.createPaymentMethodSynchronous(params)
+        assertThat(paymentMethod.type).isEqualTo(PaymentMethod.Type.Sunbit)
+    }
+
+    @Test
     fun createPaymentMethod_withMultibanco_shouldCreateObject() {
         val params = PaymentMethodCreateParamsFixtures.MULTIBANCO
         val stripe = Stripe(context, ApiKeyFixtures.MULTIBANCO_PUBLISHABLE_KEY)